requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap5==21.3
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
html5lib==1.1
idna==3.3
mechanize==0.4.7
mysqlclient==2.1.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
